revert tile size type change
Use double instead of int for tile size
so that fractional offset math works.

Convert tile identity to a double point
when computing tile offsets.

+++ lib/src/grid/constants.dart
@@ -1,4 +1,4 @@
 import 'package:flutter_map/plugin_api.dart';
 
-const _tileSize = 256;
-const tileSize = CustomPoint<int>(_tileSize, _tileSize);
+const _tileSize = 256.0;
+const tileSize = CustomPoint<double>(_tileSize, _tileSize);
